+++ src/main/java/com/service/UserService.java
@@ -10,11 +10,11 @@ import java.util.List;
 public interface UserService {
     public List<User> getAll();
 
-    public User getOne(Long id);
+    public User getOne(int id);
     @Transactional
     public void createUser(User user);
     @Transactional
-    public void deleteUser(long id);
+    public void deleteUser(int id);
     @Transactional
-    public void update(User updatedUser, Long id);
+    public void update(User updatedUser, int id);
 }
