+++ src/main/java/com/service/UserServiceImp.java
@@ -4,7 +4,6 @@ import com.model.User;
 import com.userDao.UserDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -19,7 +18,7 @@ public class UserServiceImp implements UserService {
         return resultList;
     }
 
-    public User getOne(Long id) {
+    public User getOne(int id) {
         User user = userDao.getOne(id);
         return user;
     }
@@ -30,12 +29,12 @@ public class UserServiceImp implements UserService {
     }
 
 
-    public void deleteUser(long id) {
+    public void deleteUser(int id) {
         userDao.deleteUser(id);
     }
 
 
-    public void update(User updatedUser, Long id) {
+    public void update(User updatedUser, int id) {
         userDao.update(updatedUser, id);
     }
 }
