+++ src/main/java/com/userDao/UserDao.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface UserDao {
     public List<User> getAll();
 
-    public User getOne(Long id);
+    public User getOne(int id);
 
     public void createUser(User user);
 
-    public void deleteUser(long id);
+    public void deleteUser(int id);
 
-    public void update(User updatedUser, Long id);
+    public void update(User updatedUser, int id);
 }
